Create a component for the root of the app
Move the template which defines the structure of the top-level of the UI
from `app.html.jinja2` in the hypothesis/h repo into the client.

This means that the H service only needs to serve an app.html page
containing a `<app></app>` placeholder in the body which the app will
then load into.